backend/requirements.txt
fastapi
uvicorn[standard]
python-whois
pydantic
python-multipart
python-dotenv
